corrected some BME680 problems
The configuration of the BME680 sensor was done wrong with the masks, so fixed that. Also moved when the BME680 measurement process starts so the AmbaSat board doesn't brown out if the sensor heater turns on at the same time of the AmbaSat radio. Finally, started the architecture for being able to configure the BME680 settings on flight.