refactored persisted configuration code
Refactored the code pertaining to persisting a sensor's configuration in the EEPROM. Montivation was that I saw tha thte monolithic PersistedConfiguration class that knew the initimate details of all senors' configuration was not going to end well as I added more configuration  to each sensor. So the move is to make the PersistedConfiguration more of a coordinator and each Sensor class reponsibile for the configuration of that sensor, as orchestrated by PersistedConfiguration.